Simplify replacing jump labels with instruction index
The old code used a list of pairs, to maintain the mapping
from line numbers to instruction index. Since, it's a one-to-one
mapping we can use the list index, as the index for the instruction.

Assembler now also may output instruction indices which are out
of bounds. This is to cover the case when a jump is performed
to beyond the instruction memory, because the program is finished.