Use Word32 for internals of simulator, add range check for constants in assembly
The simulator now use Word32 internally, instead of Int, just like
in the MIPS architecture.

In addition, the constants in assembly are now checked for proper range.
So immediate values, must be 16-bit numbers, shamt 5 bits, address 26 bits.

Also, an error ADT is used for assembler error, instead of String.